+++ test/optimization/test_mode_estimators.jl
@@ -50,10 +50,10 @@ using Optim, OptimizationOptimJL
     end
 
 
-    @testset "ModeAsDefined" begin
+    @testset "PredefinedOptimum" begin
         context = BATContext()
-        @test @inferred(bat_findmode(prior, ModeAsDefined(), context)).result == true_mode
-        @test @inferred(bat_findmode(batmeasure(prior), ModeAsDefined(), context)).result == true_mode
+        @test @inferred(bat_findmode(prior, PredefinedOptimum(), context)).result == true_mode
+        @test @inferred(bat_findmode(batmeasure(prior), PredefinedOptimum(), context)).result == true_mode
         let post_modes = @inferred(bat_findmode(posterior, context)).result
             for k in keys(post_modes)
                 @test isapprox(post_modes[k], true_mode[k], atol=0.001)
@@ -62,10 +62,10 @@ using Optim, OptimizationOptimJL
     end
 
 
-    @testset "MaxDensitySearch" begin
+    @testset "EmpiricalOptimum" begin
         context = BATContext()
-        @test @inferred(bat_findmode(samples, MaxDensitySearch(), context)).result isa NamedTuple
-        m = bat_findmode(samples, MaxDensitySearch(), context)
+        @test @inferred(bat_findmode(samples, EmpiricalOptimum(), context)).result isa NamedTuple
+        m = bat_findmode(samples, EmpiricalOptimum(), context)
         @test samples[m.mode_idx].v == m.result
         @test isapprox(unshaped(m.result, elshape(samples.v)), true_mode_flat, rtol = 0.05)
     end
@@ -101,6 +101,10 @@ using Optim, OptimizationOptimJL
         context = BATContext(rng = Philox4x((0, 0)))
         # result is not type-stable:
         test_findmode(posterior, OptimizationAlg(optalg = OptimizationOptimJL.NelderMead(), pretransform = DoNotTransform()), 0.01, context, inferred = false) 
+
+        context = BATContext(rng = Philox4x((0, 0)), ad = ADSelector(ForwardDiff))
+        # result is not type-stable:
+        test_findmode(posterior, OptimizationAlg(optalg = Optimization.LBFGS(), trafo = DoNotTransform()), 0.01, context, inferred = false) 
     end
 
     @testset "Optimization.jl with custom options" begin # checks that options are correctly passed to Optimization.jl